+++ src/app/(auth)/login/page.tsx
@@ -1,7 +1,7 @@
 import { redirect } from "next/navigation";
 import AuthContainer from "../_components/AuthContainer";
 import LoginForm from "../_components/LoginForm";
-import { auth } from "@/app/api/auth/auth";
+import { auth } from "@/auth";
 
 export default async function LoginPage() {
   const session = await auth();
