+++ src/app/(auth)/signup/page.tsx
@@ -1,7 +1,7 @@
 import { redirect } from "next/navigation";
 import AuthContainer from "../_components/AuthContainer";
 import SignupForm from "../_components/SignupForm";
-import { auth } from "@/app/api/auth/auth";
+import { auth } from "@/auth";
 
 export default async function SignupPage() {
   const session = await auth();
