+++ src/middleware.ts
@@ -1,4 +1,4 @@
-import { auth } from "./app/api/auth/auth";
+import { auth } from "./auth";
 
 export default auth((req) => {
   if (!req.auth && req.nextUrl.pathname !== "/login") {
